Move cli files to cmd/
