+++ main.go
@@ -7,13 +7,13 @@ import (
 	"os/signal"
 	"syscall"
 
-	"github.com/HallyG/fingrab/cmd/fingrab/root"
+	"github.com/HallyG/fingrab/cmd"
 )
 
 func main() {
 	ctx, _ := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
 
-	if err := root.Main(ctx, os.Args, os.Stdout); err != nil {
+	if err := cmd.Main(ctx, os.Args, os.Stderr); err != nil {
 		fmt.Fprintf(os.Stderr, "Error: %s\n", err)
 		os.Exit(1)
 	}
